requirements.txt
-e .
pytest
aws-cdk.aws-s3
aws-cdk.aws-s3-notifications
aws-cdk.aws-lambda
aws-cdk.aws-lambda-event-sources
aws-cdk.aws-sqs
aws-cdk.aws-dynamodb
aws-cdk.aws-iam